Restructure files for css
Css files now have their own folder: /src/css/
